requirements.txt
Django==5.1.2
psycopg==3.2.3
psycopg2==2.9.9
sqlparse==0.5.1
djangorestframework==3.15.2
typing_extensions==4.12.2
